Extract deepEqual tests to own file
